+++ tests/dummy/app/Http/Controllers/TagsController.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace DummyApp\Http\Controllers;
-
-use CloudCreativity\LaravelJsonApi\Http\Controllers\JsonApiController;
-
-class TagsController extends JsonApiController
-{
-}