+++ tests/dummy/app/Http/Controllers/VideosController.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace DummyApp\Http\Controllers;
-
-use CloudCreativity\LaravelJsonApi\Http\Controllers\JsonApiController;
-
-class VideosController extends JsonApiController
-{
-}